Move MockRssShuffleManager to GetShuffleReportForMultiPartTest & Fix reversePartitionToServers
